services/stt-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
faster-whisper==0.10.0
numpy==1.24.3
scipy==1.11.4
torch==2.1.2
torchaudio==2.1.2
pyaudio==0.2.14
webrtcvad==2.0.10
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
